+++ Scroller/Scroller.js
@@ -17,435 +17,19 @@
  */
 
 import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
-import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
-import {getRect} from '@enact/spotlight/src/utils';
 import {ResizeContext} from '@enact/ui/Resizable';
-import ri from '@enact/ui/resolution';
-import utilDOM from '@enact/ui/useScroll/utilDOM';
 import {ScrollerBasic as UiScrollerBasic} from '@enact/ui/Scroller';
 import PropTypes from 'prop-types';
-import React, {Component, useCallback, useEffect} from 'react';
+import React from 'react';
 
-import useScroll from '../Scrollable';
-import Scrollbar from '../Scrollable/Scrollbar';
+import useScroll from '../useScroll';
+import Scrollbar from '../useScroll/Scrollbar';
 import Skinnable from '../Skinnable';
 
-import {useEventKey} from './useEvent';
-import {useSpotlightConfig} from './useSpotlight';
+import ScrollerBasic from './ScrollerBasic';
+import useThemeScroller from './useThemeScroller';
 
-const dataContainerDisabledAttribute = 'data-spotlight-container-disabled';
-/**
- * An Agate-styled base component for [Scroller]{@link agate/Scroller.Scroller}.
- * In most circumstances, you will want to use the
- * [SpotlightContainerDecorator]{@link spotlight/SpotlightContainerDecorator.SpotlightContainerDecorator}
- * and the Scrollable version, [Scroller]{@link agate/Scroller.Scroller}.
- *
- * @class ScrollerBasic
- * @memberof agate/Scroller
- * @extends ui/Scroller.ScrollerBasic
- * @ui
- * @public
- */
-class ScrollerBasic extends Component {
-	static displayName = 'ScrollerBasic'
-
-	static propTypes = /** @lends agate/Scroller.ScrollerBasic.prototype */ {
-		/**
-		 * Passes the instance of [Scroller]{@link ui/Scroller.Scroller}.
-		 *
-		 * @type {Object}
-		 * @param {Object} ref
-		 * @private
-		 */
-		initUiChildRef: PropTypes.func,
-
-		/**
-		 * Called when [Scroller]{@link agate/Scroller.Scroller} updates.
-		 *
-		 * @type {function}
-		 * @private
-		 */
-		onUpdate: PropTypes.func,
-
-		/**
-		 * `true` if rtl, `false` if ltr.
-		 *
-		 * @type {Boolean}
-		 * @private
-		 */
-		rtl: PropTypes.bool,
-
-		/**
-		 * Called when [Scroller]{@link agate/Scroller.Scroller} should be scrolled
-		 * and the focus should be moved to a scrollbar button.
-		 *
-		 * @type {function}
-		 * @private
-		 */
-		scrollAndFocusScrollbarButton: PropTypes.func,
-
-		/**
-		 * The spotlight id for the component.
-		 *
-		 * @type {String}
-		 * @private
-		 */
-		spotlightId: PropTypes.string
-	}
-}
-
-const useSpottable = (props, instances) => {
-	const {scrollContentHandle, scrollContentRef} = instances;
-
-	// Hooks
-
-	useSpotlightConfig(props, instances);
-
-	const {addGlobalKeyDownEventListener, removeGlobalKeyDownEventListener} = useEventKey();
-
-	const setContainerDisabled = useCallback((bool) => {
-		const
-			{spotlightId} = props,
-			containerNode = document.querySelector(`[data-spotlight-id="${spotlightId}"]`);
-
-		if (containerNode) {
-			containerNode.setAttribute(dataContainerDisabledAttribute, bool);
-
-			if (bool) {
-				addGlobalKeyDownEventListener(() => {
-					setContainerDisabled(false);
-				});
-			} else {
-				removeGlobalKeyDownEventListener();
-			}
-		}
-	}, [addGlobalKeyDownEventListener, props, removeGlobalKeyDownEventListener]);
-
-	useEffect(() => {
-		return () => setContainerDisabled(false);
-	}, [setContainerDisabled]);
-
-	useEffect(() => {
-		const {onUpdate} = props;
-
-		if (onUpdate) {
-			onUpdate();
-		}
-	});
-
-	// Functions
-
-	/**
-	 * Returns the first spotlight container between `node` and the scroller
-	 *
-	 * @param {Node} node A DOM node
-	 *
-	 * @returns {Node|Null} Spotlight container for `node`
-	 * @private
-	 */
-	function getSpotlightContainerForNode (node) {
-		do {
-			if (node.dataset.spotlightId && node.dataset.spotlightContainer && !node.dataset.expandableContainer) {
-				return node;
-			}
-		} while ((node = node.parentNode) && node !== scrollContentRef.current);
-	}
-
-	/**
-	 * Calculates the "focus bounds" of a node. If the node is within a spotlight container, that
-	 * container is scrolled into view rather than just the element.
-	 *
-	 * @param {Node} node Focused node
-	 *
-	 * @returns {Object} Bounds as returned by `getBoundingClientRect`
-	 * @private
-	 */
-	function getFocusedItemBounds (node) {
-		node = getSpotlightContainerForNode(node) || node;
-		return node.getBoundingClientRect();
-	}
-
-	/**
-	 * Calculates the new `scrollTop`.
-	 *
-	 * @param {Node} focusedItem node
-	 * @param {Number} itemTop of the focusedItem / focusedContainer
-	 * @param {Number} itemHeight of focusedItem / focusedContainer
-	 * @param {Object} scrollInfo position info. Uses `scrollInfo.previousScrollHeight`
-	 * and `scrollInfo.scrollTop`
-	 * @param {Number} scrollPosition last target position, passed scroll animation is ongoing
-	 *
-	 * @returns {Number} Calculated `scrollTop`
-	 * @private
-	 */
-	function calculateScrollTop (item) {
-		const threshold = ri.scale(48);
-		const roundToStart = (sb, st) => {
-			// round to start
-			if (st < threshold) return 0;
-
-			return st;
-		};
-		const roundToEnd = (sb, st, sh) => {
-			// round to end
-			if (sh - (st + sb.height) < threshold) return sh - sb.height;
-
-			return st;
-		};
-		// adding threshold into these determinations ensures that items that are within that are
-		// near the bounds of the scroller cause the edge to be scrolled into view even when the
-		// itme itself is in view (e.g. due to margins)
-		const isItemBeforeView = (ib, sb, d) => ib.top + d - threshold < sb.top;
-		const isItemAfterView = (ib, sb, d) => ib.top + d + ib.height + threshold > sb.top + sb.height;
-		const canItemFit = (ib, sb) => ib.height <= sb.height;
-		const calcItemAtStart = (ib, sb, st, d) => ib.top + st + d - sb.top;
-		const calcItemAtEnd = (ib, sb, st, d) => ib.top + ib.height + st + d - (sb.top + sb.height);
-		const calcItemInView = (ib, sb, st, sh, d) => {
-			if (isItemBeforeView(ib, sb, d)) {
-				return roundToStart(sb, calcItemAtStart(ib, sb, st, d));
-			} else if (isItemAfterView(ib, sb, d)) {
-				return roundToEnd(sb, calcItemAtEnd(ib, sb, st, d), sh);
-			}
-			return st;
-		};
-
-		const container = getSpotlightContainerForNode(item);
-		const scrollerBounds = scrollContentRef.current.getBoundingClientRect();
-		let {scrollHeight, scrollTop} = scrollContentRef.current;
-		let scrollTopDelta = 0;
-
-		const adjustScrollTop = (v) => {
-			scrollTopDelta = scrollTop - v;
-			scrollTop = v;
-		};
-
-		if (container) {
-			const containerBounds = container.getBoundingClientRect();
-
-			// if the entire container fits in the scroller, scroll it into view
-			if (canItemFit(containerBounds, scrollerBounds)) {
-				return calcItemInView(containerBounds, scrollerBounds, scrollTop, scrollHeight, scrollTopDelta);
-			}
-
-			// if the container doesn't fit, adjust the scroll top ...
-			if (containerBounds.top > scrollerBounds.top) {
-				// ... to the top of the container if the top is below the top of the scroller
-				adjustScrollTop(calcItemAtStart(containerBounds, scrollerBounds, scrollTop, scrollTopDelta));
-			}
-			// removing support for "snap to bottom" for 2.2.8
-			// } else if (containerBounds.top + containerBounds.height < scrollerBounds.top + scrollerBounds.height) {
-			// 	// ... to the bottom of the container if the bottom is above the bottom of the
-			// 	// scroller
-			// 	adjustScrollTop(calcItemAtEnd(containerBounds, scrollerBounds, scrollTop, scrollTopDelta));
-			// }
-
-			// N.B. if the container covers the scrollable area (its top is above the top of the
-			// scroller and its bottom is below the bottom of the scroller), we need not adjust the
-			// scroller to ensure the container is wholly in view.
-		}
-
-		const itemBounds = item.getBoundingClientRect();
-
-		return calcItemInView(itemBounds, scrollerBounds, scrollTop, scrollHeight, scrollTopDelta);
-	}
-
-	/**
-	 * Calculates the new `scrollLeft`.
-	 *
-	 * @param {Node} focusedItem node
-	 * @param {Number} scrollPosition last target position, passed when scroll animation is ongoing
-	 *
-	 * @returns {Number} Calculated `scrollLeft`
-	 * @private
-	 */
-	function calculateScrollLeft (item, scrollPosition) {
-		const childContainerNode = scrollContentRef.current;
-		const {
-			left: itemLeft,
-			width: itemWidth
-		} = getFocusedItemBounds(item);
-
-		const
-			{rtl} = props,
-			{clientWidth} = scrollContentHandle.current.scrollBounds,
-			rtlDirection = rtl ? -1 : 1,
-			{left: containerLeft} = childContainerNode.getBoundingClientRect(),
-			scrollLastPosition = scrollPosition ? scrollPosition : scrollContentHandle.current.scrollPos.left,
-			currentScrollLeft = rtl ? (scrollContentHandle.current.scrollBounds.maxLeft - scrollLastPosition) : scrollLastPosition,
-			// calculation based on client position
-			newItemLeft = childContainerNode.scrollLeft + (itemLeft - containerLeft);
-		let nextScrollLeft = scrollContentHandle.current.scrollPos.left;
-
-		if (newItemLeft + itemWidth > (clientWidth + currentScrollLeft) && itemWidth < clientWidth) {
-			// If focus is moved to an element outside of view area (to the right), scroller will move
-			// to the right just enough to show the current `focusedItem`. This does not apply to
-			// `focusedItem` that has a width that is bigger than `scrollBounds.clientWidth`.
-			nextScrollLeft += rtlDirection * ((newItemLeft + itemWidth) - (clientWidth + currentScrollLeft));
-		} else if (newItemLeft < currentScrollLeft) {
-			// If focus is outside of the view area to the left, move scroller to the left accordingly.
-			nextScrollLeft += rtlDirection * (newItemLeft - currentScrollLeft);
-		}
-
-		return nextScrollLeft;
-	}
-
-	/**
-	 * Calculates the new top and left position for scroller based on focusedItem.
-	 *
-	 * @param {Node} item node
-	 * @param {Object} scrollInfo position info. `calculateScrollTop` uses
-	 * `scrollInfo.previousScrollHeight` and `scrollInfo.scrollTop`
-	 * @param {Number} scrollPosition last target position, passed scroll animation is ongoing
-	 *
-	 * @returns {Object} with keys {top, left} containing calculated top and left positions for scroll.
-	 * @private
-	 */
-	function calculatePositionOnFocus ({item, scrollPosition}) {
-		const containerNode = scrollContentRef.current;
-		const horizontal = scrollContentHandle.current.isHorizontal();
-		const vertical = scrollContentHandle.current.isVertical();
-
-		if (!vertical && !horizontal || !item || !utilDOM.containsDangerously(containerNode, item)) {
-			return;
-		}
-
-		const containerRect = getRect(containerNode);
-		const itemRect = getRect(item);
-
-		if (horizontal && !(itemRect.left >= containerRect.left && itemRect.right <= containerRect.right)) {
-			scrollContentHandle.current.scrollPos.left = calculateScrollLeft(item, scrollPosition);
-		}
-
-		if (vertical && !(itemRect.top >= containerRect.top && itemRect.bottom <= containerRect.bottom)) {
-			scrollContentHandle.current.scrollPos.top = calculateScrollTop(item);
-		}
-
-		return scrollContentHandle.current.scrollPos;
-	}
-
-	function focusOnNode (node) {
-		if (node) {
-			Spotlight.focus(node);
-		}
-	}
-
-	// Return
-
-	return {
-		calculatePositionOnFocus,
-		focusOnNode,
-		setContainerDisabled
-	};
-};
-
-const useThemeScroller = (props) => {
-	const {scrollContentHandle, scrollContentRef} = props;
-
-	// Hooks
-
-	const {calculatePositionOnFocus, focusOnNode, setContainerDisabled} = useSpottable(props, {scrollContentHandle, scrollContentRef});
-
-	useEffect(() => {
-		props.setThemeScrollContentHandle({
-			calculatePositionOnFocus,
-			focusOnNode,
-			setContainerDisabled
-		});
-	}, [calculatePositionOnFocus, focusOnNode, props, props.setThemeScrollContentHandle, setContainerDisabled]);
-
-	// Render
-
-	const propsObject = Object.assign({}, props);
-
-	delete propsObject.scrollContainerContainsDangerously;
-	delete propsObject.onUpdate;
-	delete propsObject.scrollAndFocusScrollbarButton;
-	delete propsObject.setThemeScrollContentHandle;
-	delete propsObject.spotlightId;
-	delete propsObject.scrollContainerHandle;
-
-	return propsObject;
-};
-
-/**
- * Allows 5-way navigation to the scrollbar controls. By default, 5-way will
- * not move focus to the scrollbar controls.
- *
- * @name focusableScrollbar
- * @memberof agate/Scroller.ScrollerBasic.prototype
- * @type {Boolean}
- * @default false
- * @public
- */
-
-/**
- * Unique identifier for the component.
- *
- * When defined and when the `Scroller` is within a [Panel]{@link agate/Panels.Panel}, the
- * `Scroller` will store its scroll position and restore that position when returning to the
- * `Panel`.
- *
- * @name id
- * @memberof agate/Scroller.ScrollerBasic.prototype
- * @type {String}
- * @public
- */
-
-/**
- * Sets the hint string read when focusing the next button in the vertical scroll bar.
- *
- * @name scrollDownAriaLabel
- * @memberof agate/Scroller.ScrollerBasic.prototype
- * @type {String}
- * @default $L('scroll down')
- * @public
- */
-
-/**
- * Sets the hint string read when focusing the previous button in the horizontal scroll bar.
- *
- * @name scrollLeftAriaLabel
- * @memberof agate/Scroller.ScrollerBasic.prototype
- * @type {String}
- * @default $L('scroll left')
- * @public
- */
-
-/**
- * Sets the hint string read when focusing the next button in the horizontal scroll bar.
- *
- * @name scrollRightAriaLabel
- * @memberof agate/Scroller.ScrollerBasic.prototype
- * @type {String}
- * @default $L('scroll right')
- * @public
- */
-
-/**
- * Sets the hint string read when focusing the previous button in the vertical scroll bar.
- *
- * @name scrollUpAriaLabel
- * @memberof agate/Scroller.ScrollerBasic.prototype
- * @type {String}
- * @default $L('scroll up')
- * @public
- */
-
-/**
- * An Agate-styled Scroller, Scrollable applied.
- *
- * Usage:
- * ```
- * <Scroller>Scroll me.</Scroller>
- * ```
- *
- * @class Scroller
- * @memberof agate/Scroller
- * @extends agate/Scroller.ScrollerBasic
- * @ui
- * @public
- */
 let Scroller = (props) => {
 	// Hooks
 
@@ -548,5 +132,5 @@ Scroller = Skinnable(
 export default Scroller;
 export {
 	Scroller,
-	ScrollerBasic
+	ScrollerBasic as ScrollerBase
 };
