+++ useScroll/ScrollButtons.js
@@ -32,17 +32,17 @@ const
 	};
 
 /**
- * An Agate-styled scroll buttons. It is used in [Scrollbar]{@link agate/Scrollable.Scrollbar}.
+ * An Agate-styled scroll buttons. It is used in [Scrollbar]{@link agate/useScroll.Scrollbar}.
  *
  * @class ScrollButtons
- * @memberof agate/Scrollable
+ * @memberof agate/useScroll
  * @ui
  * @private
  */
 class ScrollButtons extends Component {
 	static displayName = 'ScrollButtons'
 
-	static propTypes = /** @lends agate/Scrollable.ScrollButtons.prototype */ {
+	static propTypes = /** @lends agate/useScroll.ScrollButtons.prototype */ {
 		/**
 		 * The render function for thumb.
 		 *
