+++ useScroll/useScroll.js
@@ -1,11 +1,3 @@
-/**
- * Provides Agate-themed scrollable components and behaviors.
- *
- * @module agate/Scrollable
- * @exports Scrollable
- * @private
- */
-
 import {forward} from '@enact/core/handle';
 import platform from '@enact/core/platform';
 import Spotlight from '@enact/spotlight';
@@ -17,12 +9,12 @@ import {useScrollContentHandle} from '@enact/ui/useScroll/useScrollContentHandle
 import {assignPropertiesOf} from '@enact/ui/useScroll';
 import utilDOM from '@enact/ui/useScroll/utilDOM';
 import utilEvent from '@enact/ui/useScroll/utilEvent';
-import PropTypes from 'prop-types';
-import {Component, useContext, useRef} from 'react';
+import {useContext, useRef} from 'react';
 
 import $L from '../internal/$L';
 import {SharedState} from '../Panels/SharedStateDecorator';
 
+import ScrollableBasic from './ScrollableBasic';
 import {useThemeScrollContentHandle} from './useThemeScrollContentHandle';
 import {
 	useEventFocus, useEventKey, useEventMonitor, useEventMouse,
@@ -43,7 +35,7 @@ const
  * [VirtualGridList]{@link agate/VirtualList.VirtualGridList}.
  *
  * @constant dataIndexAttribute
- * @memberof agate/Scrollable
+ * @memberof agate/useScroll
  * @type {String}
  * @private
  */
@@ -55,154 +47,6 @@ const getTargetInViewByDirectionFromPosition = (direction, position, container) 
 	const target = getTargetByDirectionFromPosition(direction, position, Spotlight.getActiveContainer());
 	return getIntersectingElement(target, container);
 };
-
-/**
- * An Agate-styled component that provides horizontal and vertical scrollbars.
- *
- * @class ScrollableBase
- * @memberof agate/Scrollable
- * @extends ui/Scrollable.ScrollableBase
- * @ui
- * @public
- */
-class ScrollableBase extends Component { // ScrollableBase is now only used in storybook.
-	static displayName = 'Scrollable'
-
-	static propTypes = /** @lends agate/Scrollable.Scrollable.prototype */ {
-		/**
-		 * Render function.
-		 *
-		 * @type {Function}
-		 * @required
-		 * @private
-		 */
-		childRenderer: PropTypes.func.isRequired,
-
-		/**
-		 * This is set to `true` by SpotlightContainerDecorator
-		 *
-		 * @type {Boolean}
-		 * @private
-		 */
-		'data-spotlight-container': PropTypes.bool,
-
-		/**
-		 * `false` if the content of the list or the scroller could get focus
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @private
-		 */
-		'data-spotlight-container-disabled': PropTypes.bool,
-
-		/**
-		 * This is passed onto the wrapped component to allow
-		 * it to customize the spotlight container for its use case.
-		 *
-		 * @type {String}
-		 * @private
-		 */
-		'data-spotlight-id': PropTypes.string,
-
-		/**
-		 * Direction of the list or the scroller.
-		 * `'both'` could be only used for[Scroller]{@link agate/Scroller.Scroller}.
-		 *
-		 * Valid values are:
-		 * * `'both'`,
-		 * * `'horizontal'`, and
-		 * * `'vertical'`.
-		 *
-		 * @type {String}
-		 * @private
-		 */
-		direction: PropTypes.oneOf(['both', 'horizontal', 'vertical']),
-
-		/**
-		 * Allows 5-way navigation to the scrollbar controls. By default, 5-way will
-		 * not move focus to the scrollbar controls.
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @public
-		 */
-		focusableScrollbar: PropTypes.bool,
-
-		/**
-		 * A unique identifier for the scrollable component.
-		 *
-		 * When specified and when the scrollable is within a SharedStateDecorator, the scroll
-		 * position will be shared and restored on mount if the component is destroyed and
-		 * recreated.
-		 *
-		 * @type {String}
-		 * @public
-		 */
-		id: PropTypes.string,
-
-		/**
-		 * Specifies preventing keydown events from bubbling up to applications.
-		 * Valid values are `'none'`, and `'programmatic'`.
-		 *
-		 * When it is `'none'`, every keydown event is bubbled.
-		 * When it is `'programmatic'`, an event bubbling is not allowed for a keydown input
-		 * which invokes programmatic spotlight moving.
-		 *
-		 * @type {String}
-		 * @default 'none'
-		 * @private
-		 */
-		preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
-
-		/**
-		 * Sets the hint string read when focusing the next button in the vertical scroll bar.
-		 *
-		 * @type {String}
-		 * @default $L('scroll down')
-		 * @public
-		 */
-		scrollDownAriaLabel: PropTypes.string,
-
-		/**
-		 * Sets the hint string read when focusing the previous button in the horizontal scroll bar.
-		 *
-		 * @type {String}
-		 * @default $L('scroll left')
-		 * @public
-		 */
-		scrollLeftAriaLabel: PropTypes.string,
-
-		/**
-		 * Sets the hint string read when focusing the next button in the horizontal scroll bar.
-		 *
-		 * @type {String}
-		 * @default $L('scroll right')
-		 * @public
-		 */
-		scrollRightAriaLabel: PropTypes.string,
-
-		/**
-		 * Sets the hint string read when focusing the previous button in the vertical scroll bar.
-		 *
-		 * @type {String}
-		 * @default $L('scroll up')
-		 * @public
-		 */
-		scrollUpAriaLabel: PropTypes.string,
-
-		/*
-		 * TBD
-		 */
-		type: PropTypes.string
-	}
-
-	static defaultProps = {
-		'data-spotlight-container-disabled': false,
-		focusableScrollbar: false,
-		preventBubblingOnKeyDown: 'none',
-		type: 'JS'
-	}
-}
 
 const useThemeScroll = (props, instances, context) => {
 	const {themeScrollContentHandle, scrollContentRef, scrollContainerHandle, scrollContainerRef} = instances;
@@ -590,21 +434,9 @@ const useScroll = (props) => {
 	};
 };
 
-/**
- * An Agate-styled component that provides horizontal and vertical scrollbars.
- *
- * @class Scrollable
- * @memberof agate/Scrollable
- * @mixes spotlight/SpotlightContainerDecorator
- * @extends agate/Scrollable.ScrollableBase
- * @ui
- * @public
- */
-
 export default useScroll;
 export {
 	dataIndexAttribute,
-	ScrollableBase as Scrollable,
-	ScrollableBase,
+	ScrollableBasic as ScrollableBase,
 	useScroll
 };
